+++ src/structures/paging/mod.rs
@@ -10,7 +10,6 @@ pub use self::recursive::*;
 use core::fmt;
 use core::marker::PhantomData;
 use core::ops::{Add, AddAssign, Sub, SubAssign};
-use os_bootinfo;
 use ux::*;
 use {PhysAddr, VirtAddr};
 
@@ -427,24 +426,6 @@ impl<S: PageSize> Iterator for PhysFrameRange<S> {
         } else {
             None
         }
-    }
-}
-
-impl From<os_bootinfo::FrameRange> for PhysFrameRange {
-    fn from(range: os_bootinfo::FrameRange) -> Self {
-        PhysFrameRange {
-            start: PhysFrame::from_start_address(PhysAddr::new(range.start_addr())).unwrap(),
-            end: PhysFrame::from_start_address(PhysAddr::new(range.end_addr())).unwrap(),
-        }
-    }
-}
-
-impl Into<os_bootinfo::FrameRange> for PhysFrameRange {
-    fn into(self) -> os_bootinfo::FrameRange {
-        os_bootinfo::FrameRange::new(
-            self.start.start_address().as_u64(),
-            self.end.start_address().as_u64(),
-        )
     }
 }